backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
passlib
bcrypt==3.2.0
python-jose[cryptography]
python-multipart
python-dotenv
email-validator
